Added Exchange Sort
Added exchange sort to algorithms.cpp and Sorting.h, The exchange sort
compares each element of an array and swap those elements that are not
in their proper position, just like a bubble sort does. The only
difference between the two sorting algorithms is the manner in which
they compare the elements.

The exchange sort compares the first element with each element of the
array, making a swap where is necessary.